api_yamdb/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt
django-filter==2.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
asgiref==3.2.10
pytz==2020.1
sqlparse==0.3.1 
